backend/pyproject.toml
[project]
name = "open-experiment"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.117.1",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.25",
    #
    # Open Telemetry Requirements
    #
    "opentelemetry-distro>=0.58b0",
    "opentelemetry-exporter-otlp>=1.37.0",
    # Generated by running command in README.md
    "opentelemetry-instrumentation-asyncio==0.58b0",
    "opentelemetry-instrumentation-click==0.58b0",
    "opentelemetry-instrumentation-dbapi==0.58b0",
    "opentelemetry-instrumentation-fastapi==0.58b0",
    "opentelemetry-instrumentation-httpx==0.58b0",
    "opentelemetry-instrumentation-jinja2==0.58b0",
    "opentelemetry-instrumentation-logging==0.58b0",
    "opentelemetry-instrumentation-sqlalchemy==0.58b0",
    "opentelemetry-instrumentation-sqlite3==0.58b0",
    "opentelemetry-instrumentation-starlette==0.58b0",
    "opentelemetry-instrumentation-threading==0.58b0",
    "opentelemetry-instrumentation-tortoiseorm==0.58b0",
    "opentelemetry-instrumentation-urllib==0.58b0",
    "opentelemetry-instrumentation-urllib3==0.58b0",
    "opentelemetry-instrumentation-wsgi==0.58b0",
    #
    # Open Feature requirements
    #
    "typing-extensions>=4.15.0",
    "openfeature-sdk>=0.8.3",
    "openfeature-provider-flagsmith>=0.1.5",
    "flagsmith>=3.10.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "requests>=2.32.5",
    "ruff>=0.13.2",
]
